Work around HHVM bug by using XMLWriter::writeAttribute()
HHVM has a bug which causes a null namespace URI passed to
writeAttributeNS() to be converted to an empty string before being
passed to libxml2. This causes libxml2 to output an invalid namespace
declaration. So instead, use writeAttribute() to write the prefixed
attribute name, bypassing libxml2's namespace support.

libxml2 keeps a stack of current xmlns declarations, so you would
think it would be possible to declare a namespace on the root element
and then use it in subsequent elements by specifying the same URI each
time. But libxml does not parse xmlns:prefix="..." attributes to add
them to its declaration list, and has no other way to declare
namespaces, so the relevant attributes would have to actually be present
on the root element.

Bug: T117534
Change-Id: Iae8c95adb7824dd9d44849dd54f1ccc67e70e084